CDAP-14738 python improvements move out commons
